Rename tree.h to btree.h.
